# Dogfood new_uninit and maybe_uninit_slice in rustc_arena